[souschef] Extract FP16
This will extract external FP16 to separate header.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
